Add original dataset.
This was done without knowing which pieces of DNA had vector
attached or not to them; we've revamped the library prep to get
a piece of the vector in each read so we can better call integration
sites that have actual vector attached to them.

The prediction is this will lower the false positive rate by a ton.